scripts/requirements.txt
requests>=2.31.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0



